PageContent or beta information and components, removing static content
